requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
psycopg2-binary
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
gunicorn
boto3
